requirements.txt
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.9.0.80
Pillow==11.0.0
pytesseract==0.3.10
